+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Fixed
 - Some style issues.
 - Description in `list` field is not shown [#30].
+- GitHub adapter doesn't show files bigger than 1Mb [#31].
 
 ## [0.7.0] - 2024-10-11
 ### Added
@@ -437,6 +438,7 @@ First version
 [#28]: https://github.com/lumeland/cms/issues/28
 [#29]: https://github.com/lumeland/cms/issues/29
 [#30]: https://github.com/lumeland/cms/issues/30
+[#31]: https://github.com/lumeland/cms/issues/31
 
 [0.7.1]: https://github.com/lumeland/cms/compare/v0.7.0...HEAD
 [0.7.0]: https://github.com/lumeland/cms/compare/v0.6.8...v0.7.0