+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## [0.7.1] - Unreleased
+## [0.7.1] - 2024-10-31
 ### Fixed
 - Some style issues.
 - Description in `list` field is not shown [#30].
@@ -440,7 +440,7 @@ First version
 [#30]: https://github.com/lumeland/cms/issues/30
 [#31]: https://github.com/lumeland/cms/issues/31
 
-[0.7.1]: https://github.com/lumeland/cms/compare/v0.7.0...HEAD
+[0.7.1]: https://github.com/lumeland/cms/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/lumeland/cms/compare/v0.6.8...v0.7.0
 [0.6.8]: https://github.com/lumeland/cms/compare/v0.6.7...v0.6.8
 [0.6.7]: https://github.com/lumeland/cms/compare/v0.6.6...v0.6.7